requirements.txt
fastapi
uvicorn
pydantic
numpy
tensorflow
scipy
scikit-learn
joblib
psycopg2-binary
